Fix minor issues

+++ Habitica/src/main/java/com/habitrpg/android/habitica/data/implementation/UserRepositoryImpl.kt
@@ -64,7 +64,7 @@ class UserRepositoryImpl(localRepository: UserLocalRepository, apiClient: ApiCli
                         val calendar = GregorianCalendar()
                         val timeZone = calendar.timeZone
                         val offset = -TimeUnit.MINUTES.convert(timeZone.getOffset(calendar.timeInMillis).toLong(), TimeUnit.MILLISECONDS)
-                        if (offset != user.preferences?.timezoneOffset ?: 0) {
+                        if (offset.toInt() != user.preferences?.timezoneOffset ?: 0) {
                             return@flatMap updateUser(user, "preferences.timezoneOffset", offset.toString())
                         } else {
                             return@flatMap Flowable.just(user)
